feat:  adding keys to ssh-agent works!
I think the flow I want to follow is relying on ssh-agent and
ssh_config instead of hooking up a terminal instance for a SSH
connection.
I'm thinking about the times when we need to `scp` or `rsync` with
these boxes and having a terminal isn't going to help. `sb ssh` should
get you authed to do whatever you need to do using the tools you
already use.